% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{add_cleavages}
\alias{add_cleavages}
\title{add_cleavages}
\usage{
add_cleavages(joined_with_library, n_residues = 4)
}
\arguments{
\item{joined_with_library}{a tibble containing columns named "peptide",
"library_match_sequence", and "library_real_sequence".}

\item{n_residues}{the number of residues to the left and right of the
cleavage site to include in the output.}
}
\value{
a tibble with cleavage information added.
}
\description{
Adds cleavage information to a tibble by wraping the n_term_cleavage
and c_term_cleavage functions into a consolidated function.
}
\keyword{internal}
