% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_helper_functions.R
\name{add_peptide_data}
\alias{add_peptide_data}
\title{add_peptide_data}
\usage{
add_peptide_data(tibble, qf)
}
\arguments{
\item{tibble}{tibble you would like to add peptide info to. Must have column 
named peptide}

\item{qf}{a QFeatures object with rowData for peptides. cleavage_seq,
cleavage_pos, and cleavage_type.}
}
\value{
a tibble with column named peptide.
}
\description{
adds peptide information for every peptide in the data.
}
\keyword{internal}
