% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_possible_8mers_from_228_library}
\alias{all_possible_8mers_from_228_library}
\title{all_possible_8mers_from_228_library
All possible 8mers from the standard (as of 26April2024)
228 MSP-MS peptide library
(This is equivalent to the result of
mspms::calculate_all_cleavages(mspms::peptide_library$real_cleavage_seq,n=4))
vector of the 14 AA peptides used in the library.}
\format{
## `all_possible_8mers_from_228_library`
A vector with 2964 entries
}
\source{
<standard peptide library used with MSP-MS method
in the O’Donoghue lab as of 26April2024>
}
\usage{
all_possible_8mers_from_228_library
}
\description{
all_possible_8mers_from_228_library
All possible 8mers from the standard (as of 26April2024)
228 MSP-MS peptide library
(This is equivalent to the result of
mspms::calculate_all_cleavages(mspms::peptide_library$real_cleavage_seq,n=4))
vector of the 14 AA peptides used in the library.
}
\keyword{datasets}
