% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceLogo_helper_fuctions.R
\name{calc_AA_count_of_motif}
\alias{calc_AA_count_of_motif}
\title{calc_AA_count_of_motif}
\usage{
calc_AA_count_of_motif(cleavage_motif)
}
\arguments{
\item{cleavage_motif}{a vector of cleavage motifs}
}
\value{
a matrix of counts
}
\description{
Calculate the counts of amino acids at each position of a motif for all the
sequences in a vector.
}
\keyword{internal}
