% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AA_fc.R
\name{calc_AA_fc}
\alias{calc_AA_fc}
\title{calc_AA_fc}
\usage{
calc_AA_fc(experimental_prop_matrix, background_prop_matrix)
}
\arguments{
\item{experimental_prop_matrix}{this is a matrix of the experimental proportions (from your vector of cleavage sequences) at each position.}

\item{background_prop_matrix}{this is a matrix of the background proportions of AAs at each position}
}
\value{
a data frame.
}
\description{
calculate the fold change of each AA by position.
}
\examples{
experimental_prop_matrix = matrix(c(0.4,0.2),nrow = 1,dimnames = list("A"))
background_prop_matrix = matrix(c(0.2,0.4),nrow = 1,dimnames = list("A"))
calc_AA_fc(experimental_prop_matrix, background_prop_matrix)




}
