% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceLogo_helper_fuctions.R
\name{calc_AA_motif_zscore}
\alias{calc_AA_motif_zscore}
\title{calc_AA_motif_zscore}
\usage{
calc_AA_motif_zscore(
  background_count_matrix,
  background_prop_matrix,
  experimental_count_matrix,
  experimental_prop_matrix
)
}
\arguments{
\item{background_count_matrix}{the count matrix from the
background sequences}

\item{background_prop_matrix}{the proportion matrix from the
background sequences}

\item{experimental_count_matrix}{the count matrix from the
experimental sequences}

\item{experimental_prop_matrix}{the proportion matrix from the
experimental sequences}
}
\value{
a data frame of Zscores for each amino acid at each position.
}
\description{
Calculate the Z score for the amino acids at each position
}
\keyword{internal}
