% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceLogo_helper_fuctions.R
\name{calc_AA_percent_difference}
\alias{calc_AA_percent_difference}
\title{calc_AA_percent_difference}
\usage{
calc_AA_percent_difference(background_prop_matrix, experimental_prop_matrix)
}
\arguments{
\item{background_prop_matrix}{a proportion matrix of amino acids
per position from background cleavage sequences}

\item{experimental_prop_matrix}{a proportion matrix of amino acids per
position from experimental cleavage sequences}
}
\value{
a data frame of percent differences
}
\description{
Calculate the percent difference between a matrix of background proportions
and a matrix of experimentally observed proportions.
}
\keyword{internal}
