% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceLogo_helper_fuctions.R
\name{calc_AA_prop_of_motif}
\alias{calc_AA_prop_of_motif}
\title{calc_AA_prop_of_motif}
\usage{
calc_AA_prop_of_motif(count_matrix)
}
\arguments{
\item{count_matrix}{this is a matrix of the counts of cleavage motifs}
}
\value{
a matrix with proportions of counts.
}
\description{
calculate the proportion of amino acids at each position in a vector of
motifs.
}
\examples{

count_matrix <- matrix(c(1, 10, 100, 1000, 9, 90, 900, 9000),
  nrow = 2,
  byrow = TRUE
)
calc_AA_prop_of_motif(count_matrix)

}
\keyword{internal}
