% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bits.R
\name{calc_bits}
\alias{calc_bits}
\title{calc_bits}
\usage{
calc_bits(proportion)
}
\arguments{
\item{proportion}{a numeric value representing the proportion of a given amino acid}
}
\value{
a numeric value representing the bits of the given proportion
}
\description{
function to calculate the bits of a given proportion as described: https://iomics.ugent.be/icelogoserver/resources/manual.pdf
}
\examples{
calc_bits(0.25)
}
