% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_helper_functions.R
\name{calc_limma_design_matrix}
\alias{calc_limma_design_matrix}
\title{calc_limma_design_matrix}
\usage{
calc_limma_design_matrix(colData, norm_data)
}
\arguments{
\item{colData}{colData with condition and time variables as factors}
}
\value{
a model matrix
}
\description{
Calculates a limma compatible design matrix for mspms data.
}
\keyword{internal}
