% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helper_functions.R
\name{calc_per_samples_library_nd}
\alias{calc_per_samples_library_nd}
\title{calc_per_samples_library_nd
Calculate the percentage of samples each library_id peptide was not detected
in.}
\usage{
calc_per_samples_library_nd(
  processed_qf,
  peptide_library_ids = mspms::peptide_library$library_id
)
}
\arguments{
\item{processed_qf}{a QFeatures object with a SummarizedExperiment named
"peptides". Intended to be prepared by one of the pre-processing
prepare_x_data functions of the mspms R package.}

\item{peptide_library_ids}{a character vector containing the names of the
library_ids}
}
\value{
a tibble containing percentage of samples each library id was
detected in,  both as full length, and as cleavage products.
}
\description{
calc_per_samples_library_nd
Calculate the percentage of samples each library_id peptide was not detected
in.
}
\keyword{internal}
