% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceLogo_helper_fuctions.R
\name{calc_sig_zscores}
\alias{calc_sig_zscores}
\title{calc_sig_zscores
Determine which Zscores are significant at the given alpha for a
matrix of scores}
\usage{
calc_sig_zscores(zscores, pval = 0.05)
}
\arguments{
\item{zscores}{= a data frame of zscores}

\item{pval}{= p value threshold for significance. Default is 0.05}
}
\value{
a tibble of significant zscores
}
\description{
calc_sig_zscores
Determine which Zscores are significant at the given alpha for a
matrix of scores
}
\keyword{internal}
