% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{check_file_is_valid_diann}
\alias{check_file_is_valid_diann}
\title{check_file_is_valid_diann
Check to make sure the input data looks like the expected DIA-NN output
file.}
\usage{
check_file_is_valid_diann(diann_data)
}
\arguments{
\item{diann_data}{pg_matrix.tsv file generated by DIA-NN and read
into R.}
}
\value{
a stop command with a informative message if file looks unexpected.
otherwise, nothing.
}
\description{
check_file_is_valid_diann
Check to make sure the input data looks like the expected DIA-NN output
file.
}
