% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{check_file_is_valid_fragpipe}
\alias{check_file_is_valid_fragpipe}
\title{check_file_is_valid_fragpipe
Check to make sure the input data looks like the expected FragPipe file.}
\usage{
check_file_is_valid_fragpipe(fragpipe_data)
}
\arguments{
\item{fragpipe_data}{combined_peptide.tsv file generated by FragPipe read
into R.}
}
\value{
a stop command with a informative message if file looks unexpected.
otherwise, nothing.
}
\description{
check_file_is_valid_fragpipe
Check to make sure the input data looks like the expected FragPipe file.
}
