% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{check_file_is_valid_pd}
\alias{check_file_is_valid_pd}
\title{check_file_is_valid_pd
Check to make sure the input data looks like the expected ProteomeDiscoverer
file.}
\usage{
check_file_is_valid_pd(pd_data)
}
\arguments{
\item{pd_data}{PeptideGroups.txt file generated by ProteomeDiscover and read
into R.}
}
\value{
a stop command with a informative message if file looks unexpected.
otherwise, nothing.
}
\description{
check_file_is_valid_pd
Check to make sure the input data looks like the expected ProteomeDiscoverer
file.
}
