% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{check_file_is_valid_peaks}
\alias{check_file_is_valid_peaks}
\title{check_file_is_valid_peaks
check to make sure a loaded tibble is what a peaks file is expected to look
like.}
\usage{
check_file_is_valid_peaks(peaks_data)
}
\arguments{
\item{peaks_data}{}
}
\description{
check_file_is_valid_peaks
check to make sure a loaded tibble is what a peaks file is expected to look
like.
}
\keyword{internal}
