% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{check_file_is_valid_sage}
\alias{check_file_is_valid_sage}
\title{check_file_is_valid_sage
Check to make sure the input data looks like the expected PEAKS file.}
\usage{
check_file_is_valid_sage(sage_data)
}
\arguments{
\item{sage_data}{read in lfq.tsv file output prodcued by Sage
into R.}
}
\value{
a stop command with a informative message if file looks unexpected.
otherwise, nothing.
}
\description{
check_file_is_valid_sage
Check to make sure the input data looks like the expected PEAKS file.
}
\keyword{internal}
