% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{check_peptide_library}
\alias{check_peptide_library}
\title{check_peptide_library}
\usage{
check_peptide_library(peptide_library)
}
\arguments{
\item{peptide_library}{}
}
\value{
an informative error if the column names of the peptide library are
unexpected. Otherwise nothing.
}
\description{
check_peptide_library
}
\keyword{internal}
