% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cleavage_added_data}
\alias{cleavage_added_data}
\title{data with cleavage information added by the mspms package}
\format{
## `cleavage_added_data`
A tibble with 820 rows and 32 columns:
\describe{
  \item{Peptide}{Peptide Sequence Detected}
  \item{library_reference_id}{reference id of the detected peptide to that of the library reported by software}
  \item{library_match_sequence}{the sequence match to the peptide library}
  \item{library_real_sequence}{the real sequence of the peptide in the library}
  ...
}
}
\source{
<mspms processed data originally from PEAKS files found in "tests/testdata/protein-peptides-id.csv" and "tests/testdata/protein-peptides-lfq.csv">
}
\usage{
cleavage_added_data
}
\description{
data prepared by the mspms package with cleavage information added as described in the readme
}
\keyword{datasets}
