% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{colData}
\alias{colData}
\title{colData
A tibble containing the colData associated with an experiment to proc}
\format{
## `colData`
A tibble: 42 × 4
}
\source{
colData corresponding to cathepsin A-D MSP-MS experiment
}
\usage{
colData
}
\description{
colData
A tibble containing the colData associated with an experiment to proc
}
\keyword{datasets}
