% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{consolidate_cleavages}
\alias{consolidate_cleavages}
\title{consolidate_cleavages}
\usage{
consolidate_cleavages(cleavage_added_data)
}
\arguments{
\item{cleavage_added_data}{a tibble where cleavage information has
been added by add_cleavages()}
}
\value{
a tibble with the cleavage information combined into a single column
 and rows with no cleavage information or double information removed.
}
\description{
Consolidate the n term and c term cleavage data. The nterm and cterm
cleavage information  are consolidated into a single column and rows
}
\keyword{internal}
