% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helper_functions.R
\name{count_cleavages_per_pos}
\alias{count_cleavages_per_pos}
\title{count_cleavages_per_pos}
\usage{
count_cleavages_per_pos(data, peptide_library = mspms::peptide_library)
}
\arguments{
\item{data}{a tibble containing columns named peptide,cleavage_pos,condition,
and time. Other column names can be included.}
}
\value{
a ggplot2 object
}
\description{
Count the number of cleavages per position
}
\keyword{internal}
