% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_design_matrix.R
\name{extract_design_matrix}
\alias{extract_design_matrix}
\title{extract_design_matrix}
\usage{
extract_design_matrix(prepared_data)
}
\arguments{
\item{prepared_data}{= this is data that has been prepared for normalyzer analysis}
}
\value{
a data frame with the sample (complete names) and group id (samples that are treated as replicates of each other)
}
\description{
This extracts the design matrix from the column names in the data frame  Note that this assumes a very specific naming convention. 
In this naming convention, there are three _. 
The first specifies the 
The second specifics the time point
The third specifies the replicate. 
The combination of the first and second defines all of the distinct groups.
}
