% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_design_matrix_pd.R
\name{extract_design_matrix_pd}
\alias{extract_design_matrix_pd}
\title{extract_design_matrix_pd}
\usage{
extract_design_matrix_pd(filepath)
}
\arguments{
\item{filepath}{= this is the path to the proteome discoverer excel output}
}
\description{
extract a design matrix from the proteome discover column headers corresponding to samples.
}
