% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_re.R
\name{extract_re}
\alias{extract_re}
\title{extract_re}
\usage{
extract_re(icelogo_matrix, type = "positive", threshold = 0)
}
\arguments{
\item{icelogo_matrix}{= this is an icelogo matrix, as produced by 
prepare_icelogo_data()}

\item{type}{= "positive" or "negative" depending on whether you want to look
for positively or negatively enriched amino acids}

\item{threshold}{= the threshold to consider amino acids for inclusion. If 
type is positive, will return AAs with values greater than the threshold.
If type is negative, will return AAs with values less than the threshold.

If no amino acids are found greater than the threshold, a . will be returned 
in regular expression.}
}
\value{
a character vector containing a regular expression extracted from 
the icelogo matrix based on user defined parameters.
}
\description{
extract regular expressions from an icelogo matrix.
}
\examples{
icelogo_matrix = data.frame(P4 = c(10,NA,NA),
                           P3 = c(NA,10,NA),
                          P2 = c(NA,NA,10),
                          P1 = c(NA,NA,NA),
                           `P1'` = c(NA,NA,NA),
                           `P2'` = c(10,10,10),
                           `P3'` = c(10,10,NA),
                           `P4'` = c(NA,10,10))
rownames(icelogo_matrix) = c("A","B","C")
icelogo_matrix = as.matrix(icelogo_matrix)
extract_re(icelogo_matrix)

}
