% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nd_peptides.R
\name{find_nd_peptides}
\alias{find_nd_peptides}
\title{find_nd_peptides
Find library peptides that were not detected in the data.}
\usage{
find_nd_peptides(prepared_data, peptide_library, design_matrix)
}
\arguments{
\item{prepared_data}{a tibble with data prepared by the mspms package.}

\item{peptide_library}{a tibble containing the peptide library used}

\item{design_matrix}{a tibble containing the design matrix}
}
\value{
a tibble with the missing peptides for each sample
}
\description{
find_nd_peptides
Find library peptides that were not detected in the data.
}
\examples{
nd_peptides <- find_nd_peptides(
  mspms::peaks_prepared_data,
  mspms::peptide_library,
  mspms::design_matrix
)

}
