% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_report.R
\name{generate_report}
\alias{generate_report}
\title{generate_report}
\usage{
generate_report(
  prepared_data,
  design_matrix,
  peptide_library = mspms::peptide_library,
  n_residues = 4,
  outdir = getwd()
)
}
\arguments{
\item{prepared_data}{= this is the prepared data.}

\item{design_matrix}{= this is the design matrix.}

\item{peptide_library}{= this is the peptide library used in the experiment.}

\item{n_residues}{= this is the number of residues that you want to show
cleavage sequences for left and right of the cut site.}

\item{outdir}{= this is the output directory you would like to render the
report to.}
}
\value{
a knited .html report of the mspms analysis.
}
\description{
wrapper function to generate an automatic .html report of a basic mspms analysis.
}
\examples{
\dontshow{if (isTRUE(FALSE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

generate_report(
  prepared_data = mspms::peaks_prepared_data,
  design_matrix = mspms::design_matrix,
  outdir = "../Desktop/test_report"
)
\dontshow{\}) # examplesIf}
}
