% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_outliers.R
\name{handle_outliers}
\alias{handle_outliers}
\title{handle_outliers}
\usage{
handle_outliers(normalyzed_data)
}
\arguments{
\item{normalyzed_data}{= this is the normalzed data.}
}
\value{
a data frame in the long format where outliers according to a dixon test have been handled.
}
\description{
handle_outliers
}
