% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{imputed}
\alias{imputed}
\title{imputed}
\format{
## `imputed`
A tibble with 820 rows and 28 columns:
\describe{
  \item{Peptide}{Peptide Sequence Detected}
  \item{RT}{Retention Time}
  ...
}
}
\source{
<mspms processed data originally from PEAKS files found in
"tests/testdata/protein-peptides-id.csv" and
"tests/testdata/protein-peptides-lfq.csv">
}
\usage{
imputed
}
\description{
data imputed by the mspms package as described in the readme
}
\keyword{datasets}
