% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{joined_with_library}
\alias{joined_with_library}
\title{joined_with_library
this is an example dataset processed by mspms that has been joined with the peptide library}
\format{
## `joined_with_library`
A tibble with 820 rows and 30 columns:
\describe{
  \item{Peptide}{Peptide Sequence Detected}
  ...
}
}
\source{
<mspms processed data originally from PEAKS files found in "tests/testdata/protein-peptides-id.csv" and "tests/testdata/protein-peptides-lfq.csv">
}
\usage{
joined_with_library
}
\description{
joined_with_library
this is an example dataset processed by mspms that has been joined with the peptide library
}
\keyword{datasets}
