% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_stats.R
\name{limma_stats}
\alias{limma_stats}
\title{limma_stats}
\usage{
limma_stats(processed_qf)
}
\arguments{
\item{processed_qf}{mspms data in a QFeatures object.}
}
\value{
a tibble containing statistics
}
\description{
Calculates statistics for each condition relative to time 0 using limma for
differential analysis. Results are then formatted to be consistent with
results produced by other statistic approaches used in the mspms package
(log2fc_t_test).
}
\examples{
mspms_limma_results <- limma_stats(mspms::processed_qf)
}
