% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{load_colData}
\alias{load_colData}
\title{load_colData}
\usage{
load_colData(colData_filepath)
}
\arguments{
\item{colData_filepath}{filepath to .csv file containing colData.}
}
\value{
a tibble
}
\description{
load a .csv file containing sample colData.
}
\keyword{internal}
