% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{log2fc_t_test_data}
\alias{log2fc_t_test_data}
\title{log2fc_t_test_data
A tibble containing the results of t-tests and log2fc compared to time 0
14,497 × 19}
\format{
## `peaks_prepared_data`
A tibble: 14,497 × 19
}
\source{
<mspms processed data originally from PEAKS files found in
 "tests/testdata/protein-peptides-id.csv" and
 "tests/testdata/protein-peptides-lfq.csv">
}
\usage{
log2fc_t_test_data
}
\description{
log2fc_t_test_data
A tibble containing the results of t-tests and log2fc compared to time 0
14,497 × 19
}
\keyword{datasets}
