% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2fct_condition.R
\name{log2fct_condition}
\alias{log2fct_condition}
\title{log2fct_condition}
\usage{
log2fct_condition(mspms_data, at_time, ref_condition, comparison_condition)
}
\arguments{
\item{mspms_data}{processed mspms data}

\item{at_time}{the time you would like to use for comparison}

\item{ref_condition}{the reference (denominator) condition}

\item{comparison_condition}{the comparison (numerator) condition}
}
\value{
a tibble
}
\description{
preforms log2fc and t test analysis at a user specified time point for two specified conditions.
}
\examples{
log2t = log2fct_condition(mspms::mspms_data,
                         at_time = 60,
                         ref_condition = "DMSO",
                         comparison_condition = "MZB")
}
