% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2fct_time.R
\name{log2fct_time}
\alias{log2fct_time}
\title{log2fct_time}
\usage{
log2fct_time(mspms_data, within_condition, ref_time, comparison_time)
}
\arguments{
\item{mspms_data}{processed mspms data}

\item{within_condition}{the condition you would like to compare within}

\item{ref_time}{the reference (denominator) time}

\item{comparison_time}{the comparison (numerator) time}
}
\value{
a tibble
}
\description{
preforms log2fc and t test analysis for a user specified condition comparing
 two specified times.
}
\examples{

log2t <- log2fct_time(mspms::mspms_data,
  within_condition = "DMSO",
  ref_time = "0",
  comparison_time = "60"
)

}
