% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspms-package.R
\docType{package}
\name{mspms-package}
\alias{mspms}
\alias{mspms-package}
\title{mspms: Tools for the analysis of MSP-MS data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package provides functions for the analysis of data generated by the multiplex substrate profiling by mass spectrometry for proteases (MSP-MS) method. Data exported from upstream proteomics software is accepted as input and subsequently processed for analysis. Tools for statistical analysis, visualization, and interpretation of the data are provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/baynec2/mspms}
  \item Report bugs at \url{https://github.com/baynec2/mspms/issues}
}

}
\author{
\strong{Maintainer}: Charlie Bayne \email{baynec2@gmail.com} (\href{https://orcid.org/0000-0003-1870-5298}{ORCID})

}
\keyword{internal}
