% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspms.R
\name{mspms}
\alias{mspms}
\title{mspms
runs the standard mspms data analysis pipeline in it's entirety.}
\usage{
mspms(
  prepared_data,
  design_matrix,
  peptide_library = mspms::peptide_library,
  n_residues = 4,
  outdir = getwd()
)
}
\arguments{
\item{prepared_data}{= this is data that has been processed by one of the
prepare data functions (which one depends on the upstream software used.)}

\item{design_matrix}{= this is the design matrix.
This should be a data frame with a column named "sample" that contains the
names of the samples in the data,
a column named "group" that contains the group that the sample belongs to a
 column named "time" that contains the time the sample was incubated for,
and on named "condition" that contains condition corresponding to each sample
(commonly some kind of protease inhibitor).}

\item{peptide_library}{= this is the peptide library that has been used
in the experiment.}

\item{n_residues}{= this is the number of cleavages to the left and right
of the cleavage site you want to include in the cleavage motif.}

\item{outdir}{= this is the directory to write the results of normalyze to.}
}
\value{
a tibble with processed mspms data.
}
\description{
mspms
runs the standard mspms data analysis pipeline in it's entirety.
}
\examples{
\dontshow{if (isTRUE(FALSE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

out <- mspms::mspms(mspms::peaks_prepared_data, mspms::design_matrix)
\dontshow{\}) # examplesIf}
}
