% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspms_anova.R
\name{mspms_anova}
\alias{mspms_anova}
\title{mspms_anova}
\usage{
mspms_anova(mspms_data)
}
\arguments{
\item{mspms_data}{= this is the data that has been run through the mspms pipeline.}
}
\value{
a data frame with anova results for each peptide.
}
\description{
Performs anova on data prepared with prepare_for_stats(). Tests the effect of time for each condition. FDR corrected.
}
\examples{
anova_results = mspms_anova(mspms::mspms_data)
}
