% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mspms_data}
\alias{mspms_data}
\title{mspms_data
this is an example dataset processed by mspms that has gone through the
entire standar workflow}
\format{
## `mspms_data`
A tibble with 14,184 rows and 15 columns:
\describe{
  \item{Peptide}{Peptide Sequence Detected}
  ...
}
}
\source{
<mspms processed data originally from PEAKS files found in
"tests/testdata/protein-peptides-id.csv" and
"tests/testdata/protein-peptides-lfq.csv">
}
\usage{
mspms_data
}
\description{
mspms_data
this is an example dataset processed by mspms that has gone through the
entire standar workflow
}
\keyword{datasets}
