% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_helper_functions.R
\name{mspms_t_tests}
\alias{mspms_t_tests}
\title{mspms_t_tests}
\usage{
mspms_t_tests(processed_qf, reference_variable = "time", reference_value = "0")
}
\arguments{
\item{processed_qf}{a QFeatures object with a SummarizedExperiment named
"peptides_norm".}

\item{reference_variable}{the variable to used as a reference.}

\item{reference_value}{the value of the reference variable to use as
the reference}
}
\value{
a tibble with the t test statistics for each peptide within each
group with the supplied value at the supplied variable as reference.
}
\description{
performs t-tests for each peptide within each group for the user specified.
FDR adjustment is performed.
}
\keyword{internal}
