% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspms_tidy.R
\name{mspms_tidy}
\alias{mspms_tidy}
\title{mspms_tidy
Convert a SummarizedExperiment object within a QFeatures object 
into a tidy tibble.}
\usage{
mspms_tidy(processed_qf, se_name = "peptides_norm")
}
\arguments{
\item{processed_qf}{a QFeature object containing rowData and colData.}

\item{se_name}{the name of the SummarizedExperiment you would like to 
extract}
}
\value{
a tibble containing all the rowData, colData, and assay data for the
specified SummarizedExperiment.
}
\description{
mspms_tidy
Convert a SummarizedExperiment object within a QFeatures object 
into a tidy tibble.
}
\examples{
mspms_data = mspms_tidy(mspms::processed_qf)
}
