% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mspms_tidy_data}
\alias{mspms_tidy_data}
\title{mspms_tidy_data
A tibble containing tidy data derived from QFeatures object}
\format{
## `mspms_tidy_data`
A tibble:
}
\source{
processed_qf
}
\usage{
mspms_tidy_data
}
\description{
mspms_tidy_data
A tibble containing tidy data derived from QFeatures object
}
\keyword{datasets}
