% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalyze.R
\name{normalyze}
\alias{normalyze}
\title{normalyze: this function normalyzes the data using the NormalyzerDE package.}
\usage{
normalyze(prepared_data, design_matrix, outdir = getwd())
}
\arguments{
\item{prepared_data}{= this is the data that has been prepared}

\item{design_matrix}{= this is the design matrix}

\item{outdir}{= this is the directory you would like NormalyzerDE::nomralyzer to write output to.}
}
\value{
files written to the outdir specified
the normalyzed data is then read in as a data frame.
}
\description{
normalyze: this function normalyzes the data using the NormalyzerDE package.
}
