% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{normalyzed_data}
\alias{normalyzed_data}
\title{normalyzed_data}
\format{
## `normalyzed_data`
A tibble with 820 rows and 27 columns:
\describe{
  \item{Peptide}{Peptide Sequence Detected}
  \item{RT}{Retention Time}
  ...
}
}
\source{
<mspms processed data originally from PEAKS files found in
 "tests/testdata/protein-peptides-id.csv" and
  "tests/testdata/protein-peptides-lfq.csv">
}
\usage{
normalyzed_data
}
\description{
data normalyzed by the mspms package as described in the readme
}
\keyword{datasets}
