% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peaks_prepared_data}
\alias{peaks_prepared_data}
\title{peaks_prepared_data
A QFeatures object prepared from PEAKS data of cathepsin data/.}
\format{
## `peaks_prepared_data`
An instance of class QFeatures containing 1 assays:
[1] peptides: SummarizedExperiment with 2071 rows and 42 columns
\describe{
  \item{peptides}{Peptide Sequence Detected}
  ...
}
}
\source{
<mspms processed data originally from PEAKS files found in
 "tests/testdata/protein-peptides-id.csv" and
 "tests/testdata/protein-peptides-lfq.csv">
}
\usage{
peaks_prepared_data
}
\description{
peaks_prepared_data
A QFeatures object prepared from PEAKS data of cathepsin data/.
}
\keyword{datasets}
