% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peptide_library}
\alias{peptide_library}
\title{peptide_library}
\format{
## `peptide_library`
A  data frame with 228 rows and 3 columns:
\describe{
  \item{library_reference_id}{reference id of the detected peptide as
  put in upstream software}
  \item{library_match_sequence}{the sequence match to the peptide library,
   methionine is replaced with norleucine,which should function the same as
    methionine for proteases but has the same mass as L}
  \item{library_real_sequence}{Ls corresponding to norleucine are replaced
   back with n (for norleucine )}
  ...
}
}
\source{
<O’Donoghue lab as of 26April2024 >
}
\usage{
peptide_library
}
\description{
This is the 228 peptide library used by the O’Donoghue lab as of 26April2024.
}
\keyword{datasets}
