% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_icelogos.R
\name{plot_all_icelogos}
\alias{plot_all_icelogos}
\title{plot_all_icelogos}
\usage{
plot_all_icelogos(
  mspms_data,
  pval = 0.05,
  log2fc_threshold = 3,
  type = "percent_difference",
  background_universe = mspms::all_possible_8mers_from_228_library
)
}
\arguments{
\item{mspms_data}{= this is the data that has been polished
(via mspms::polish()) and is ready for plotting}

\item{pval}{= this is the pvalue threshold that you would like to use.}

\item{log2fc_threshold}{= this is the log2 fold change threshold that you 
would like to use. Can be positive or negative if desired.}

\item{type}{= this is the type of visualization you would like to perform,
can be either "percent_difference" or "fold_change".}

\item{background_universe}{= this is a list of all the possible cleavage
sequences in the peptide library used.}
}
\value{
a ggplot object that shows the motif of the cleavage sequences
}
\description{
wrapper function for plotting icelogos for each condition in an experiment.
 Here we will plot all icelogos for all peptides that were found to be
 signficant by t test at any timepoint.
}
\examples{
plot_all_icelogos(mspms::mspms_data)
}
