% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cleavage_frequencies.R
\name{plot_cleavage_frequencies}
\alias{plot_cleavage_frequencies}
\title{Plot cleavage-site amino acid frequencies for an experimental set relative
to a reference set}
\usage{
plot_cleavage_frequencies(
  cleavage_seqs,
  background_universe = mspms::all_possible_8mers_from_228_library,
  pval_threshold = 1
)
}
\arguments{
\item{cleavage_seqs}{Character vector of experimental cleavage sequences.}

\item{background_universe}{Character vector representing the reference
cleavage sequence set. Defaults to all possible 8-mers from the 228
library.}

\item{pval_threshold}{Numeric value. Residues with p-values greater than
this threshold will be omitted from the plot.}
}
\value{
A ggplot2 object showing amino acid frequencies at each motif
  position for the experimental and reference sets.
}
\description{
This function calculates amino acid frequencies at each motif position for
an experimental cleavage sequence set, compares them to a reference set,
filters residues based on a p-value threshold, and returns an icelogo-style
bar plot.
}
\examples{
reference_set = rep(c("AAAAAAAA"),10)
plot_cleavage_frequencies(reference_set)
}
