% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cleavage_motif.R
\name{plot_cleavage_motif}
\alias{plot_cleavage_motif}
\title{plot_cleavage_motifs}
\usage{
plot_cleavage_motif(
  cleavage_seqs,
  background_universe = mspms::all_possible_8mers_from_228_library
)
}
\arguments{
\item{cleavage_seqs}{= these are the cleavage sequences that are observed in the experiment}

\item{background_universe}{= this is a list of all the possible clevage sequences in the peptide library used.}
}
\value{
a ggplot object that shows the motif of the cleavage sequences
}
\description{
This function plots the cleavage motifs that were enriched relative to background in a manner similar to IceLogo.
https://iomics.ugent.be/icelogoserver/resources/manual.pdf
}
\examples{
cleavage_seqs = mspms::prepared_for_stats \%>\%
mspms::polish() \%>\%
dplyr::filter(condition == "DMSO",time == 240) \%>\%
dplyr::pull(cleavage_seq)
}
