% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cleavages_per_pos.R
\name{plot_cleavages_per_pos}
\alias{plot_cleavages_per_pos}
\title{plot_cleavages_per_pos}
\usage{
plot_cleavages_per_pos(sig_cleavage_data, ncol = 1)
}
\arguments{
\item{sig_cleavage_data}{a tibble of data of interest containing a column
labeled peptide, cleavage_seq, condition, and cleavage_pos.}

\item{ncol}{the number of columns to plot.}
}
\value{
a ggplot2 object
}
\description{
plot the number of cleavages at each
}
\examples{
# Defining the significant peptides
sig_cleavage_data <- log2fc_t_test_data \%>\%
  dplyr::filter(p.adj <= 0.05, log2fc > 3)
# Plotting
p1 <- mspms::plot_cleavages_per_pos(sig_cleavage_data)
p1
}
