% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fc_icelogo.R
\name{plot_fc_icelogo}
\alias{plot_fc_icelogo}
\title{plot_fc_icelogo}
\usage{
plot_fc_icelogo(prepared_fc_data)
}
\arguments{
\item{prepared_fc_data}{= matrix with prepared fold change data}
}
\value{
a ggplot2 object
}
\description{
plot an icelogo based on the fold change.
}
\examples{
prepared_fc_data <- data.frame(P1 = c(5, 10), P2 = c(1, -1))
rownames(prepared_fc_data) <- c("R", "K")
prepared_fc_data <- as.matrix(prepared_fc_data)
plot_fc_icelogo(prepared_fc_data)

}
