% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(prepared_for_stats, scale = "column")
}
\arguments{
\item{scale}{= how would you like the data scaled? default is none, but can be done by "row", "column", or "none"}

\item{cleavage_added_data}{= this is the data that has been processed by the add_cleavage() function.}
}
\value{
a heatmaply interactive heatmao
}
\description{
This produces a heatmaply interactive heatmap of the data.
}
