% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nd_peptides.R
\name{plot_nd_peptides}
\alias{plot_nd_peptides}
\title{plot_nd_peptides}
\usage{
plot_nd_peptides(
  processed_qf,
  peptide_library_ids = mspms::peptide_library$library_id
)
}
\arguments{
\item{processed_qf}{a QFeatures object containing a SummarizedExperiment
named "peptides"}

\item{peptide_library_ids}{a vector of all peptide library ids in the 
experiment.}
}
\value{
a ggplot2 object
}
\description{
plot the percentage of samples each peptide from library was undetected in
(if the percentage is > 0).
}
\examples{
plot_nd_peptides(mspms::processed_qf)
}
