% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pd_icelogo.R
\name{plot_pd_icelogo}
\alias{plot_pd_icelogo}
\title{plot_pd_icelogo}
\usage{
plot_pd_icelogo(prepared_sig_pdiff_data)
}
\arguments{
\item{prepared_sig_pdiff_data}{= this is the prepared significant percent difference matrix.}
}
\value{
a ggplot2 object
}
\description{
plot a icelogo of the percent difference.
}
\examples{
prepared_sig_pdiff_data <- data.frame(P1 = c(50, 10), P2 = c(-50, -10))
rownames(prepared_sig_pdiff_data) <- c("R", "K")
prepared_sig_pdiff_data <- as.matrix(prepared_sig_pdiff_data)
plot_pd_icelogo(prepared_sig_pdiff_data)
}
