% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rt_qc.R
\name{plot_rt_qc}
\alias{plot_rt_qc}
\title{plot_rt_qc
plot the retention times of detected peptides as a histogram or density plot}
\usage{
plot_rt_qc(
  prepared_data,
  design_matrix,
  type = "histogram",
  binwidth = 2,
  facet_by_group = TRUE
)
}
\arguments{
\item{prepared_data}{= data prepared by mspms package}

\item{design_matrix}{= design matrix}

\item{type}{= type of plot - histogram or density}

\item{binwidth}{= binwidth for histogram}

\item{facet_by_group}{= whether to facet by group, TRUE or FALSE}
}
\value{
a ggplot2 object
}
\description{
plot_rt_qc
plot the retention times of detected peptides as a histogram or density plot
}
\examples{
plot_rt_qc(mspms::peaks_prepared_data,
          mspms::design_matrix,
          type = "density",
          facet_by_group = TRUE)
}
