% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{prepare_file}
\alias{prepare_file}
\title{Generic preparation function for MSP-MS input files}
\usage{
prepare_file(
  filepath,
  colData_filepath,
  peptide_library = mspms::peptide_library,
  n_residues = 4,
  read_fun,
  validate_fun,
  transform_fun
)
}
\arguments{
\item{filepath}{path to the input file}

\item{colData_filepath}{path to colData CSV}

\item{peptide_library}{peptide library used in experiment}

\item{n_residues}{number of residues to include around cleavage site}

\item{read_fun}{function to read the file (e.g., read_tsv, read_csv, read_delim)}

\item{validate_fun}{function to validate the file (check_file_is_valid_*)}

\item{transform_fun}{function to transform file into standard peptide format}
}
\value{
a QFeatures object
}
\description{
Generic preparation function for MSP-MS input files
}
\keyword{internal}
