% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_for_stats.R
\name{prepare_for_stats}
\alias{prepare_for_stats}
\title{prepare_for_stats}
\usage{
prepare_for_stats(cleavage_added_data, design_matrix)
}
\arguments{
\item{cleavage_added_data}{= this is the processed data with cleavage information added. Intended to be downstream of the add_cleavages() function.}

\item{design_matrix}{= this is the design matrix with "sample",group","time","condition" columns.}
}
\value{
a data frame in long format with the design matrix combined.
}
\description{
This function prepares data for downstream statistics by transforming the data to the long format and combining with the design matrix.
}
