% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fragpipe.R
\name{prepare_fragpipe}
\alias{prepare_fragpipe}
\title{prepare_fragpipe}
\usage{
prepare_fragpipe(
  combined_peptide_filepath,
  colData,
  peptide_library = mspms::peptide_library,
  n_residues = 4
)
}
\arguments{
\item{combined_peptide_filepath}{file path the combined_peptide.tsv file
generated by FragPipe.}

\item{colData}{tibble containing colData (sample metadata). Must have
columns named "quantCols","group","condition",and "time".}

\item{peptide_library}{peptide library contained in the experiment}

\item{n_residues}{the number of amino acid residues before and after the
cleavage site to generate a cleavage seq for.}
}
\value{
a QFeatures object containing a summarizedExperiment named "peptides"
}
\description{
Prepare a label free quantification file exported from Fragpipe
for subsequent mspms analysis.
}
\examples{
colData = readr::read_csv("inst/extdata/colData.csv")
combined_peptide_filepath = "inst/extdata/fragpipe_combined_peptide.tsv"
prepared_fragpipe_data = prepare_fragpipe(combined_peptide_filepath)
}
