% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceLogo_helper_fuctions.R
\name{prepare_icelogo_data}
\alias{prepare_icelogo_data}
\title{prepare_icelogo_data}
\usage{
prepare_icelogo_data(
  cleavage_seqs,
  background_universe = mspms::all_possible_8mers_from_228_library,
  pval = 0.05,
  type = "percent_difference"
)
}
\arguments{
\item{cleavage_seqs}{the cleavage sequences that are observed in the
experiment}

\item{background_universe}{a vector of the cleavage sequences to use as
the background.}

\item{pval}{the p-value threshold to consdier}

\item{type}{the type of iceLogo calculation to perform. Accepted values are
"percent_difference" or "fold_change".}
}
\value{
a matrix of enriched amino acids per position
}
\description{
Prepare the final matrix containing iceLogo data for plotting.
}
\keyword{internal}
