% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_pd.R
\name{prepare_pd}
\alias{prepare_pd}
\title{prepare_pd}
\usage{
prepare_pd(filepath)
}
\arguments{
\item{filepath}{= this is the filepath to the proteome discoverer excel formatted file.}
}
\value{
a tibble with the data formatted for use with normalyze
}
\description{
prepare proteome discoverer output for mspms analysis.
}
\examples{
\dontshow{if (isTRUE(FALSE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
prepared_proteome_discoverer = prepare_pd("tests/testdata/proteome_discoverer_output.xlsx")
\dontshow{\}) # examplesIf}
}
