% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceLogo_helper_fuctions.R
\name{prepare_sig_p_dif}
\alias{prepare_sig_p_dif}
\title{prepare_sig_p_dif}
\usage{
prepare_sig_p_dif(percent_difference, sig_zscores)
}
\arguments{
\item{percent_difference}{a data frame containing the percent differences}

\item{sig_zscores}{a matrix of significant amino acids at each position
based on z-scores}
}
\value{
a tibble
}
\description{
Prepare significant percent difference data frame for iceLogo
}
\keyword{internal}
