% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prepared_for_stats}
\alias{prepared_for_stats}
\title{data prepared for statistics by the mspms package}
\format{
## `prepared_for_stats`
A tibble with 19,680 rows and 13 columns:
\describe{
  \item{Peptide}{Peptide Sequence Detected}
  \item{library_reference_id}{reference id of the detected peptide to that of the library reported by software}
  \item{library_match_sequence}{the sequence match to the peptide library}
  \item{library_real_sequence}{the real sequence of the peptide in the library}
  ...
}
}
\source{
<mspms processed data originally from PEAKS files found in "tests/testdata/protein-peptides-id.csv" and "tests/testdata/protein-peptides-lfq.csv">
}
\usage{
prepared_for_stats
}
\description{
data prepared for statistics by the mspms package as described in the readme
}
\keyword{datasets}
