% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{prepared_to_qf}
\alias{prepared_to_qf}
\title{convert prepared data to a QFeatures object}
\usage{
prepared_to_qf(
  prepared_data,
  colData,
  peptide_library = mspms::peptide_library,
  n_residues = 4
)
}
\arguments{
\item{prepared_data}{data prepared within one of the prepare functions}

\item{colData}{sample metadata}

\item{peptide_library}{the peptide library used.}

\item{n_residues}{the number of residues reported in the cleavage site}
}
\value{
a QFeatures object
}
\description{
convert prepared data to a QFeatures object
}
\keyword{internal}
