% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_qf.R
\name{process_qf}
\alias{process_qf}
\title{process_qf}
\usage{
process_qf(prepared_qf)
}
\arguments{
\item{prepared_qf}{this is a QFeatures object containing a 
SummarizedExperiment named "peptides"}
}
\value{
a QFeatures object containing a 
SummarizedExperiments named "peptides","peptides_log","peptides_log_norm",
"peptides_log_impute_norm",and "peptides_norm"
}
\description{
process_qf
}
\examples{
processed_qf = process_qf(mspms::peaks_prepared_data)
}
