% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_check.R
\name{qc_check}
\alias{qc_check}
\title{qc_check
Run simple quality control checks on the data. This checks to see how many
full length peptides belonging to the library were identified in the data,
as well as how many peptides were detected among all cleavage products}
\usage{
qc_check(prepared_data, peptide_library, design_matrix)
}
\arguments{
\item{prepared_data}{data prepared for analysis by the mspms package.}

\item{peptide_library}{the peptide library used in the experiment}

\item{design_matrix}{the design matrix containing experimental metadata}
}
\value{
a list with two slots, 1st slot contains a tibble with QC metrics
second list contains and interpretation. 2nd slot contains a tibble with the
missing peptides in each sample.
}
\description{
qc_check
Run simple quality control checks on the data. This checks to see how many
full length peptides belonging to the library were identified in the data,
as well as how many peptides were detected among all cleavage products
}
\examples{
qc_check_data <- qc_check(
  mspms::peaks_prepared_data,
  mspms::peptide_library,
  mspms::design_matrix
)
}
