% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_helper_functions.R
\name{remaining_cd_names}
\alias{remaining_cd_names}
\title{remaining_cd_names}
\usage{
remaining_cd_names(processed_qf, reference_variable)
}
\arguments{
\item{processed_qf}{}

\item{col_name}{}
}
\description{
remaining_cd_names
}
\keyword{internal}
