% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{transform_diann}
\alias{transform_diann}
\title{Transform DIA-NN report (pr_matrix.tsv) into standard peptide format}
\usage{
transform_diann(df, peptide_library)
}
\arguments{
\item{df}{DIA-NN pr_matrix.tsv read with read_tsv}

\item{peptide_library}{peptide library used in the experiment}
}
\value{
a tibble with columns: peptide, library_id, sample intensities
}
\description{
Transform DIA-NN report (pr_matrix.tsv) into standard peptide format
}
\keyword{internal}
