% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{transform_fragpipe}
\alias{transform_fragpipe}
\title{Transform FragPipe combined_peptide.tsv into standard peptide format}
\usage{
transform_fragpipe(df, peptide_library)
}
\arguments{
\item{df}{FragPipe combined_peptide.tsv read with read_tsv}

\item{peptide_library}{peptide library used in the experiment}
}
\value{
a tibble with columns: peptide, library_id, sample intensities
}
\description{
Transform FragPipe combined_peptide.tsv into standard peptide format
}
\keyword{internal}
