% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{transform_pd}
\alias{transform_pd}
\title{Transform Proteome Discoverer PeptideGroups.txt into standard peptide format}
\usage{
transform_pd(df, peptide_library)
}
\arguments{
\item{df}{Proteome Discoverer PeptideGroups.txt read with read_delim}

\item{peptide_library}{peptide library used in the experiment}
}
\value{
a tibble with columns: peptide, library_id, sample intensities
}
\description{
Transform Proteome Discoverer PeptideGroups.txt into standard peptide format
}
\keyword{internal}
