% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{transform_peaks}
\alias{transform_peaks}
\title{Transform PEAKS LFQ file into standard peptide format}
\usage{
transform_peaks(df, peptide_library, quality_threshold = 0.3)
}
\arguments{
\item{df}{PEAKS LFQ data read in with read_csv}

\item{peptide_library}{peptide library used in the experiment}

\item{quality_threshold}{minimum peptide quality to keep (default 0.3)}
}
\value{
a tibble with columns: peptide, library_id, sample intensities
}
\description{
Transform PEAKS LFQ file into standard peptide format
}
\keyword{internal}
