% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{transform_sage}
\alias{transform_sage}
\title{Transform Sage lfq.tsv into standard peptide format}
\usage{
transform_sage(df, peptide_library)
}
\arguments{
\item{df}{sage lfq.tsv read with read_tsv}

\item{peptide_library}{peptide library used in the experiment}
}
\value{
a tibble with columns: peptide, library_id, sample intensities
}
\description{
Transform Sage lfq.tsv into standard peptide format
}
\keyword{internal}
