% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cleavages.R
\name{add_cleavages}
\alias{add_cleavages}
\title{add_cleavages}
\usage{
add_cleavages(joined_with_library)
}
\arguments{
\item{joined_with_library}{}
}
\value{
a data frame with cleavage information added.
}
\description{
This functions adds cleavage information to the data frame that has been normalized, outlier removed, imputed, and joined with (peptide) library.
The cleavage information is represented as the 4 amino acids to the left and right of the cleavage site.
If there is no amino acid in a position of original peptide in the library that was cleaved, it is represented as an X.
This wraps the mspms::n_term_cleavage and mspms::c_term_cleavage functions into a consolidated function.
}
