% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_cleavages.R
\name{add_cleavages}
\alias{add_cleavages}
\title{add_cleavages}
\usage{
add_cleavages(joined_with_library, n_residues = 4)
}
\arguments{
\item{joined_with_library}{= this is the tibble that has been normalized, outlier removed, imputed, and joined with the library.}

\item{n_residues}{= the number of residues to the left and right of the cleavage site to include in the output.}
}
\value{
a tibble with cleavage information added.
}
\description{
This functions adds cleavage information to the tibble that has been normalized, outlier removed, imputed, and joined with (peptide) library.
The cleavage information is represented as the 4 amino acids to the left and right of the cleavage site.
If there is no amino acid in a position of original peptide in the library that was cleaved, it is represented as an X.
This wraps the mspms::n_term_cleavage and mspms::c_term_cleavage functions into a consolidated function.
}
\examples{
# adding the clevages 5 AA to the left and right of the cleavage site.
add_cleavages(mspms::joined_with_library,n_residues = 5)

}
