% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_AA_count_of_motif.R
\name{calc_AA_count_of_motif}
\alias{calc_AA_count_of_motif}
\title{calc_AA_count_of_motif}
\usage{
calc_AA_count_of_motif(cleavage_motif)
}
\arguments{
\item{cleavage_motif}{this is a vector of cleavage motifs that you would like to count the amino acids for.}
}
\value{
a matrix of counts
}
\description{
calculate the counts of ammino acids at each position of a motif.
}
\examples{
cleavage_motifs = c("XXXXXXXX","XXXXXXXX")
calc_AA_count_of_motif(cleavage_motifs)
}
