% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceLogo_helper_fuctions.R
\name{calc_AA_fc}
\alias{calc_AA_fc}
\title{calc_AA_fc}
\usage{
calc_AA_fc(experimental_prop_matrix, background_prop_matrix, sig_zscores)
}
\arguments{
\item{experimental_prop_matrix}{a matrix of the experimental
proportions (from your vector of cleavage sequences) at each position.}

\item{background_prop_matrix}{a matrix of the background proportions
of AAs at each position}

\item{sig_zscores}{a tibble of the significant zscores.}
}
\value{
a matrix
}
\description{
Calculate the fold change of each amino acid by position.
}
\keyword{internal}
