% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_helper_functions.R
\name{calc_per_samples_library_nd}
\alias{calc_per_samples_library_nd}
\title{Find library peptides that were not detected in the data.}
\usage{
calc_per_samples_library_nd(
  processed_qf,
  peptide_library_ids = mspms::peptide_library$library_id
)
}
\arguments{
\item{prepared_data}{a tibble with data prepared by the mspms package.}

\item{peptide_library}{a tibble containing the peptide library used}

\item{design_matrix}{a tibble containing the design matrix}
}
\value{
a tibble with the missing peptides for each sample
}
\description{
Find library peptides that were not detected in the data.
}
\examples{
nd_peptides <- find_nd_peptides(
  mspms::peaks_prepared_data,
  mspms::peptide_library,
  mspms::design_matrix
)

}
\keyword{internal}
