% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sig_zscores.R
\name{calc_sig_zscores}
\alias{calc_sig_zscores}
\title{calc_sig_zscores
determine which zscores are significant at the given alpha for a matrix of zscores}
\usage{
calc_sig_zscores(zscores, pval = 0.05)
}
\arguments{
\item{zscores}{= a data frame of zscores}

\item{pval}{= p value threshold for significance. Default is 0.05}
}
\value{
a tibble of significant zscores
}
\description{
calc_sig_zscores
determine which zscores are significant at the given alpha for a matrix of zscores
}
\examples{
zscores = data.frame(p1 = c(0,1,2,3,4,5),p2 = c(0,-1,-2,-3,-4,-5))
rownames(zscores) = c("A","B","C","D","E","F")#'
calc_sig_zscores(zscores)
}
