% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{check_file_is_valid_peaks}
\alias{check_file_is_valid_peaks}
\title{check_file_is_valid_peaks
Check to make sure the input data looks like the expected PEAKS file.}
\usage{
check_file_is_valid_peaks(peaks_data)
}
\arguments{
\item{peaks_data}{protein-peptides-lfq.csv file generated by PEAKS read
into R.}
}
\value{
a stop command with a informative message if file looks unexpected.
otherwise, nothing.
}
\description{
check_file_is_valid_peaks
Check to make sure the input data looks like the expected PEAKS file.
}
\keyword{internal}
