% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_cleavages_per_pos.R
\name{count_cleavages_per_pos}
\alias{count_cleavages_per_pos}
\title{count_cleavages_per_pos}
\usage{
count_cleavages_per_pos(mspms_data, sig_peptides)
}
\arguments{
\item{mspms_data}{= this is the data that has gone through the
mspms pipeline.}

\item{sig_peptides}{= this is a list of peptides of interest.}
}
\value{
a tibble with the number of cleavages per position
}
\description{
count the number of cleavages per position
}
\examples{
# Extracting the significant peptides for DMSO
sig_peptides <- mspms::mspms_anova(mspms::mspms_data) \%>\%
  dplyr::filter(
    p.adj <= 0.05,
    condition == "MZB"
  ) \%>\%
  dplyr::pull(Peptide)

# Counting the number of cleavages per position
counts <- count_cleavages_per_pos(mspms::mspms_data, sig_peptides)

# Plotting the results
p1 <- counts \%>\%
  ggplot2::ggplot(ggplot2::aes(x = cleavage_pos, y = n)) +
  ggplot2::geom_point() +
  ggplot2::geom_line()
p1

}
