% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cterm_cleavage.R
\name{cterm_cleavage}
\alias{cterm_cleavage}
\title{cterm_cleavage = finding the cleavage sequences in reference to a p}
\usage{
cterm_cleavage(peptide_sequence, library_match_sequence, library_real_sequence)
}
\arguments{
\item{peptide_sequence}{= this is the peptide sequence measured. _ denotes clevage site.}

\item{library_match_sequence}{= this is the sequence of the corresponding peptide in the library to the peptide_sequence.}

\item{library_real_sequence}{= this is the library match sequence with the Ls transformed to Ms (For some reason we don't understand.)}
}
\value{
a data frame with the peptide sequence, cleavage sequences 4 AA on the left and right of the c term cleavage, and the position of the c term cleavage in the library sequence.
}
\description{
cterm_cleavage = finding the cleavage sequences in reference to a p
}
