% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_helper_functions.R
\name{cterm_cleavage}
\alias{cterm_cleavage}
\title{cterm_cleavage}
\usage{
cterm_cleavage(
  peptide_sequence,
  library_match_sequence,
  library_real_sequence,
  n_residues = 4
)
}
\arguments{
\item{peptide_sequence}{= this is the peptide sequence in single letter
AA code. _ denotes cleavage site.}

\item{library_match_sequence}{= this is the sequence of the corresponding
peptide in the library to the peptide_sequence.}

\item{library_real_sequence}{= this is the library match sequence with the
Ls transformed to Ms.}

\item{n_residues}{= this is the number of residues to the left and right of
the cleavage site that you want to extract.}
}
\value{
a tibble with the peptide sequence, cleavage sequences n number of
AA on the left and right of the c term cleavage, and the position of the
c-term cleavage in the library sequence.
}
\description{
Finding the cleavage sequences on the C terminus of a given peptide in
reference to the peptide library it was derived from
}
\examples{
peptide_sequence <- "YWLSTHLAGKR_R"
library_match_sequence <- "YWLSTHLAGKRRDW"
library_real_sequence <- "YWMSTHLAGKRRDW"
cterm_cleavage(
  peptide_sequence, library_match_sequence,
  library_real_sequence
)

}
\keyword{internal}
