% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_design_matrix_pd.R
\name{extract_design_matrix_pd}
\alias{extract_design_matrix_pd}
\title{extract_design_matrix_pd}
\usage{
extract_design_matrix_pd(filepath)
}
\arguments{
\item{filepath}{= this is the path to the proteome discoverer excel output}
}
\value{
a tibble containing the extracted design matrix compatable with the normalyze function
}
\description{
extract a design matrix from the proteome discover column headers corresponding to samples.
}
\details{
Note this depends on how you have decided to label your samples in the proteome discoverer software.
see ../tests/testthat/test-prepare_pd.R for an example.

Abundance: F5: Sample, 1, 15 min is an example of a sample name input by the software that this is designed to work with

in this example, samples gets assigned to the pd_group, 1 gets assigned to the group, and 15 min gets assigned to the time.
}
\examples{

design_matrix = extract_design_matrix_pd("tests/testdata/proteome_discoverer_output.xlsx")
}
