% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2fc_t_test.R
\name{log2fc_t_test}
\alias{log2fc_t_test}
\title{log2fc_t_test}
\usage{
log2fc_t_test(mspms_data)
}
\arguments{
\item{mspms_data}{= this is data that has been run in the mspms pipeline.}
}
\value{
a tibble containing log2fc and t test statistics across conditions
}
\description{
Gives log2fc and statistics from t tests across conditions.
 Gives differences relative to time 0 within each condition.
}
\examples{

log2fc_and_t_test <- log2fc_t_test(mspms::mspms_data)

}
