% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspms-package.R
\docType{package}
\name{mspms-package}
\alias{mspms}
\alias{mspms-package}
\title{mspms: Analysis tools for use with MSP-MS}
\description{
This package provides functions for the analysis of data generated by the Multiplex substrate profiling by mass spectrometry for proteases (MSP-MS) method.
}
\keyword{internal}
