% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspms_anova.R
\name{mspms_anova}
\alias{mspms_anova}
\title{mspms_anova}
\usage{
mspms_anova(prepared_for_stats)
}
\arguments{
\item{prepared_for_stats}{= this is the data that has been prepared using prepare_for_stats()}
}
\value{
a data frame with anova results for each peptide.
}
\description{
Performs anova on data prepared with prepare_for_stats(). Tests the effect of time for each condition. FDR corrected.
}
