% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspms_log2fc.R
\name{mspms_log2fc}
\alias{mspms_log2fc}
\title{mspms_log2fc}
\usage{
mspms_log2fc(prepared_for_stats)
}
\arguments{
\item{prepared_for_stats}{= this is the data that has been prepared for stats using the prepare for stats function.}
}
\value{
a data frame with the log2fc for each timepoint within each condition relative to time 0
}
\description{
Calculates the log2fc for each timepoint within each condition relative to time 0.
}
