% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspms_log2fc.R
\name{mspms_log2fc}
\alias{mspms_log2fc}
\title{mspms_log2fc}
\usage{
mspms_log2fc(mspms_data)
}
\arguments{
\item{mspms_data}{= this is the data that has been run through the mspms
pipeline}
}
\value{
a data frame with the log2fc for each timepoint within each condition
relative to time 0
}
\description{
Calculates the log2fc for each timepoint within each condition relative to
time 0.
}
\examples{

log2fc <- mspms_log2fc(mspms::prepared_for_stats)
}
