% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mspms_t_tests.R
\name{mspms_t_tests}
\alias{mspms_t_tests}
\title{mspms_t_tests}
\usage{
mspms_t_tests(prepared_for_stats)
}
\arguments{
\item{prepared_for_stats}{= this is the data that has been prepared for stats using prepare_for_stats()}
}
\value{
a data frame with the t test statistics for each peptide within each group with t0 as reference
}
\description{
Performs ttests for each peptide within each group with t0 as reference. P value adjustment performed using FDR correction
}
