% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalyze.R
\name{normalyze}
\alias{normalyze}
\title{normalyze: This function performs normalization of MSP-MS data.}
\usage{
normalyze(prepared_data, design_matrix, outdir = getwd())
}
\arguments{
\item{prepared_data}{= this is the data that has been prepared via
mspms::prepare_for_normalyzer().}

\item{design_matrix}{= this is the design matrix. This should be a data frame
 with a column named "sample" that contains the names of the samples
 in the data,
a column named "group" that contains the group that the sample belongs to a
column named "time" that contains the time the sample was incubated for,
and on named "condition" that contains condition cooresponding to each sample
 (commonly some kind of protease inhibitor).}

\item{outdir}{= this is the directory you would like NormalyzerDE::nomralyzer
to write output to.Default is current working dir.}
}
\value{
Files from NormalyzerDE are written to the outdir specified.

The normalized data is returned as a tibble.
}
\description{
This function uses the NormalyzerDE package to perform normalization.
It returns the median normalized data as a data frame.
For this normalization the Intensity of each variable in a given sample is
 divided by the median of intensities of all variables
in the sample and then multiplied by the mean of median of sum of
intensities of all variables in all samples.
}
\examples{
\dontshow{if (isTRUE(FALSE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

normalyzed_data <- normalyze(mspms::peaks_prepared_data, mspms::design_matrix,
  outdir = getwd()
)
\dontshow{\}) # examplesIf}
}
