% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nterm_cleavage.R
\name{nterm_cleavage}
\alias{nterm_cleavage}
\title{nterm_cleavage}
\usage{
nterm_cleavage(peptide_sequence, library_match_sequence, library_real_sequence)
}
\arguments{
\item{peptide_sequence}{= this is the peptide sequence in single leter AA code. _ denotes cleavage site..}

\item{library_match_sequence}{= this is the sequence of the corresponding peptide in the library to the peptide_sequence.}

\item{library_real_sequence}{= this is the library match sequence with the Ls transformed to Ms (This is what the legacy code did so it is kept this way in case there was a good reason for it)}
}
\value{
a data frame with the peptide sequence, cleavage sequences 4 AA on the left and right of the n term cleavage, and the position of the n term cleavage in the library sequence.
}
\description{
Finding the cleavage sequences on the N terminus of a given peptide in reference to the peptide library it was derived from.
}
\examples{

peptide_sequence = "F_LAHWVGI"
library_real_sequence = "GMYYKRFMAHWVGI"
library_match_sequence = "GLYYKRFLAHWVGI"


nterm_cleavage(peptide_sequence,library_real_sequence,library_match_sequence)
}
