% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nterm_cleavage.R
\name{nterm_cleavage}
\alias{nterm_cleavage}
\title{nterm_cleavage:}
\usage{
nterm_cleavage(peptide_sequence, library_match_sequence, library_real_sequence)
}
\arguments{
\item{peptide_sequence}{= this is the peptide sequence measured. _ denotes clevage site.}

\item{library_match_sequence}{= this is the sequence of the corresponding peptide in the library to the peptide_sequence.}

\item{library_real_sequence}{= this is the library match sequence with the Ls transformed to Ms (For some reason we don't understand.)}
}
\value{
a data frame with the peptide sequence, cleavage sequences 4 AA on the left and right of the n term cleavage, and the position of the n term cleavage in the library sequence.
}
\description{
nterm_cleavage:
}
\examples{

peptide_sequence = "F_LAHWVGI"
library_real_sequence = "GMYYKRFMAHWVGI"
library_match_sequence = "GLYYKRFLAHWVGI"


nterm_cleavage(peptide_sequence,library_real_sequence,library_match_sequence)
}
