% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{peaks_prepared_data}
\alias{peaks_prepared_data}
\title{peaks_prepared_data
this is an example dataset from peaksthat has been processed by mspms and is
 ready for normalyzation.}
\format{
## `peaks_prepared_data`
A tibble with 820 rows and 27 columns:
\describe{
  \item{Peptide}{Peptide Sequence Detected}
  ...
}
}
\source{
<mspms processed data originally from PEAKS files found in
 "tests/testdata/protein-peptides-id.csv" and
 "tests/testdata/protein-peptides-lfq.csv">
}
\usage{
peaks_prepared_data
}
\description{
peaks_prepared_data
this is an example dataset from peaksthat has been processed by mspms and is
 ready for normalyzation.
}
\keyword{datasets}
