% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cleavages_per_pos.R
\name{plot_cleavages_per_pos}
\alias{plot_cleavages_per_pos}
\title{plot_cleavages_per_pos}
\usage{
plot_cleavages_per_pos(count_of_cleavages, facets = "condition", ncol = 1)
}
\arguments{
\item{count_of_cleavages}{= tibble with number of cleavages per position, for
each condition and time point}

\item{facets}{= condition, grid, or wrap If grid facets for each condition and 
time are shown even if data does not exist. 
Wrap on the other hand is condensed. None shows plots overlayed by time with
different colors.}

\item{ncol}{= number of columns to plot}
}
\value{
a ggplot2 object
}
\description{
make a plot of the count of cleavages per position.
}
\examples{
c = data.frame(time = rep(60,13),
                  condition = rep("DMSO",13),
                  cleavage_pos = 1:13,
                  n = 1:13)
plot_cleavages_per_pos(c)
}
