% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(processed_qf, scale = "column", plot_method = "plotly")
}
\arguments{
\item{processed_qf}{a QFeatures object containing a SummarizedExperiment
named "peptides_norm"}

\item{scale}{how would you like the data scaled? default is none,
but can also be "row", "column", or "none"}

\item{plot_method}{what plot method would you like to use, can use
plotly or ggplot2.}
}
\value{
a heatmaply interactive heatmap
}
\description{
This produces a heatmaply interactive heatmap of the QFeatures object with 
color bars representing the condition and time for each sample in each row.
}
\details{
Each column has a colored bar representing whether the peptide is a 
cleavage product or a full length member of the peptide library.
}
\examples{
plot_heatmap(mspms::processed_qf)
}
