% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{plot_heatmap}
\usage{
plot_heatmap(mspms_data, scale = "column", plot_method = "plotly")
}
\arguments{
\item{mspms_data}{= this is the data that has been processed by the
mspms pipeline.}

\item{scale}{= how would you like the data scaled? default is none,
but can be done by "row", "column", or "none"}

\item{plot_method}{= interactive plotly heatmap or static ggplot2 heatmap}
}
\value{
a heatmaply interactive heatmap
}
\description{
This produces a heatmaply interactive heatmap of the data with color bars
 representing the condition and time for each sample.
}
\examples{
\dontshow{if (isTRUE(FALSE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

plot_heatmap(mspms::mspms_data, scale = "column")
\dontshow{\}) # examplesIf}
}
