% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_icelogo.R
\name{plot_icelogo}
\alias{plot_icelogo}
\title{plot_icelogo}
\usage{
plot_icelogo(
  cleavage_seqs,
  background_universe = mspms::all_possible_8mers_from_228_library,
  type = "percent_difference"
)
}
\arguments{
\item{cleavage_seqs}{= these are the cleavage sequences that are observed in the experiment}

\item{background_universe}{= this is a list of all the possible clevage sequences in the peptide library used.}

\item{type}{= this is the type of visualization you would like to perform, can be either "percent_difference" or "fold_change".}
}
\value{
a ggplot object that shows the motif of the cleavage sequences
}
\description{
This function plots the cleavage motifs that were enriched relative to background in a manner similar to IceLogo.
https://iomics.ugent.be/icelogoserver/resources/manual.pdf
}
\examples{
cleavage_seqs = mspms::prepared_for_stats \%>\%
mspms::polish() \%>\%
dplyr::filter(condition == "DMSO",time == 240) \%>\%
dplyr::pull(cleavage_seq)
plot_icelogo(cleavage_seqs)
}
