% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nd_peptides.R
\name{plot_nd_peptides}
\alias{plot_nd_peptides}
\title{plot_nd_peptides}
\usage{
plot_nd_peptides(nd_peptides)
}
\arguments{
\item{nd_peptides}{a tibble containing info on peptides not detected by the 
find_nd_peptides function.}
}
\value{
a ggplot2 object
}
\description{
plot peptides from libray that were not detected in the data
}
\examples{
nd_peptides <- find_nd_peptides(
  mspms::peaks_prepared_data,
  mspms::peptide_library,
  mspms::design_matrix
)
plot_nd_peptides(nd_peptides)
}
