% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{plot_pca}
\usage{
plot_pca(prepared_for_stats)
}
\arguments{
\item{prepared_for_stats}{= this is the data that has been prepared for stats by the prepare_for_stats() function}
}
\value{
a ggplot2 object
}
\description{
this creates a pca plot of the data where the points are colored by time and shaped by condition. elipses are drawn around the points at a 95% confidence interval.
}
\examples{

plot_pca(mspms::prepared_for_stats)


}
