% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{plot_pca}
\usage{
plot_pca(
  mspms_tidy_data,
  value_colname = "peptides_norm",
  color = "time",
  shape = "condition"
)
}
\arguments{
\item{mspms_tidy_data}{tidy mspms data (prepared from QFeatures object
by mspms_tidy)}

\item{value_colname}{the name of the column containing values.}

\item{color}{the name of the variable you would like to color by.}

\item{shape}{the name of the variable that you would like to determine
shape by.}
}
\value{
a ggplot2 object
}
\description{
Easily create a PCA plot from a QFeatures object containing mspms data.
Ellipses are drawn around the points at a 95% confidence interval.
Shape and colors are user specified.
}
\examples{
plot_pca(mspms::mspms_tidy_data)
}
