% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pca.R
\name{plot_pca}
\alias{plot_pca}
\title{plot_pca}
\usage{
plot_pca(mspms_data, color = "time", shape = "condition")
}
\arguments{
\item{mspms_data}{= this is the data that has been run through the mspms pipeline.}

\item{color}{= this is the string name of the variable you would like to color by.}

\item{shape}{= this is the string name of the vairable that you would like to determine shape by.}
}
\value{
a ggplot2 object
}
\description{
this creates a pca plot of the data where the points are colored by time and shaped by condition. elipses are drawn around the points at a 95% confidence interval.
}
\examples{

plot_pca(mspms::mspms_data)

}
