% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qc_check.R
\name{plot_qc_check}
\alias{plot_qc_check}
\title{plot_qc_check
plot the qc results generated by the qc_check function to show metrics about
the percent of the library detected per sample group.}
\usage{
plot_qc_check(qc_check_data, thresold_percent = 90, ncol = 1)
}
\arguments{
\item{qc_check_data}{qc check results produced by the qc_check function}

\item{ncol}{n columns.}

\item{threshold_percent}{percent to use as threshold visualized as red
dashed line}
}
\value{
a list containing 2 ggplot objects. The first slot contains plot of
percent of the full length peptides detected per sample group, the second
slot contains the percent of library mapped to per sample group (includes
full length and cleaved peptides)
}
\description{
plot_qc_check
plot the qc results generated by the qc_check function to show metrics about
the percent of the library detected per sample group.
}
\examples{
qc_check <- qc_check(
  mspms::prepared_data,
  mspms::peptide_library,
  mspms::design_matrix
)

plot_qc_check(qc_check)
}
