% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_course.R
\name{plot_time_course}
\alias{plot_time_course}
\title{plot_time_course}
\usage{
plot_time_course(mspms_data)
}
\arguments{
\item{mspms_data}{this is data from the mspms pipeline.
It should have the following columns: time, value, condition, and Peptide. It is in long format.
Can be filtered to only include the peptides you want to plot.}
}
\value{
a ggplot2 object
}
\description{
this function plots the time course of each peptide in the data set.
}
\examples{

p1 = mspms::mspms_data \%>\%
dplyr::filter(Peptide == "A_GLFNYNQLRGF") \%>\%
mspms::plot_time_course()

p1
}
