% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_volcano.R
\name{plot_volcano}
\alias{plot_volcano}
\title{plot_volcano}
\usage{
plot_volcano(log2fc_t_test_data, log2fc_threshold = 3, facets = "grid")
}
\arguments{
\item{log2fc_t_test_data}{= this is the data frame that contains the log2fc
and adjusted p values}

\item{log2fc_threshold}{= this is the threshold that you want displayed
on plot}

\item{facets}{= grid or wrap. If grid facets for each condition and time are
shown even if data does not exist. Wrap on the other hand is condensed.}
}
\value{
a ggplot2 object
}
\description{
create a volcano plot to generate log2fc and adjusted p values for
experimental conditions
}
\examples{
# Generating log2fc t test data
log2fc_t_test_data <- mspms::log2fc_t_test(mspms::mspms_data)
# plotting data.
p1 <- mspms::plot_volcano(log2fc_t_test_data, log2fc_threshold = 3)
p1
}
