% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_fc.R
\name{prepare_fc}
\alias{prepare_fc}
\title{prepare_fc}
\usage{
prepare_fc(fold_change, sig_zscores)
}
\arguments{
\item{fold_change}{= this is a matrix of the fold changes of the AA by
position.}

\item{sig_zscores}{= this is a tibble of the significant zscores.}
}
\value{
a matrix of the fold changes of the significant AAs at each position.
}
\description{
prepare fold changes of AA by position for Icelogo visualization.
}
\examples{
\dontshow{if (isTRUE(FALSE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
