% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iceLogo_helper_fuctions.R
\name{prepare_icelogo_data}
\alias{prepare_icelogo_data}
\title{prepare_icelogo_data}
\usage{
prepare_icelogo_data(
  cleavage_seqs,
  background_universe = mspms::all_possible_8mers_from_228_library,
  pval = 0.05,
  type = "percent_difference"
)
}
\arguments{
\item{cleavage_seqs}{= these are the cleavage sequences that are observed in the experiment}

\item{background_universe}{= this is a list of all the possible cleavage sequences in the peptide library used.}

\item{pval}{= this is the pvalue threshold that you would like to use.}

\item{type}{= this is the type of Icelogo matrix you would like to prepare, can be either "percent_difference" or "fold_change".}
}
\value{
a matrix of enriched amino acids per position
}
\description{
prepare a matrix containing icelogo data.
}
\examples{
cleavage_seqs = unique(mspms_data$cleavage_seq[1:100])
prepare_icelogo_data(cleavage_seqs)

}
\keyword{internal}
