% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_pd.R
\name{prepare_pd}
\alias{prepare_pd}
\title{prepare_pd}
\usage{
prepare_pd(filepath)
}
\arguments{
\item{filepath}{= this is the filepath to the proteome discoverer excel formatted file.}
}
\description{
prepare proteome discoverer output for mspms analysis.
}
