% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_peaks.R
\name{prepare_peaks}
\alias{prepare_peaks}
\title{prepare_peaks
This function prepares the data out from peaks for downstream normalization
via normalyzer.
Briefly, it combines the two files by peptide IDs and filters out peptides
that have quality scores <= 0.3. 0s are also replaced with NAs}
\usage{
prepare_peaks(lfq_filepath, id_filepath)
}
\arguments{
\item{lfq_filepath}{= this is the filepath to the first PEAKS output table}

\item{id_filepath}{= this is the filepath to the second PEAKS output table}
}
\value{
a tibble containing the combined columns from the lfq and id files,
 with quality scores > 0.3 and 0s replaced with NAs.
}
\description{
prepare_peaks
This function prepares the data out from peaks for downstream normalization
via normalyzer.
Briefly, it combines the two files by peptide IDs and filters out peptides
that have quality scores <= 0.3. 0s are also replaced with NAs
}
\examples{
\dontshow{if (isTRUE(FALSE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

prepare_peaks(
  "tests/testdata/protein-peptides-lfq.csv",
  "tests/testdata/protein-peptides-id.csv"
)
\dontshow{\}) # examplesIf}
}
