% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_helper_functions.R
\name{prepare_qc_check_data}
\alias{prepare_qc_check_data}
\title{prepare_qc_check
Run simple quality control checks on the data. This checks to see how many
full length peptides belonging to the library were identified in the data,
as well as how many peptides were detected among all cleavage products}
\usage{
prepare_qc_check_data(processed_qf, peptide_library = mspms::peptide_library)
}
\arguments{
\item{peptide_library}{the peptide library used in the experiment}

\item{prepared_data}{data prepared for analysis by the mspms package.}

\item{design_matrix}{the design matrix containing experimental metadata}
}
\value{
a list with two slots, 1st slot contains a tibble with QC metrics
second list contains and interpretation. 2nd slot contains a tibble with the
missing peptides in each sample.
}
\description{
prepare_qc_check
Run simple quality control checks on the data. This checks to see how many
full length peptides belonging to the library were identified in the data,
as well as how many peptides were detected among all cleavage products
}
\examples{
qc_check_data <- qc_check(
  mspms::peaks_prepared_data,
  mspms::peptide_library,
  mspms::design_matrix
)
}
\keyword{internal}
