% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_sig_p_dif.R
\name{prepare_sig_p_dif}
\alias{prepare_sig_p_dif}
\title{prepare_sig_pdif}
\usage{
prepare_sig_p_dif(percent_difference, sig_zscores)
}
\arguments{
\item{percent_difference}{= data frame containing the percent differences}

\item{sig_zscores}{= matrix of significant AAs at each position based on
z-scores}
}
\value{
a tibble
}
\description{
prepare significant percent difference data frame for ice logo
}
\examples{
\dontshow{if (isTRUE(FALSE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
