% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics_helper_functions.R
\name{remaining_cd_names}
\alias{remaining_cd_names}
\title{remaining_cd_names}
\usage{
remaining_cd_names(processed_qf, reference_variable)
}
\arguments{
\item{processed_qf}{a QFeatures object}

\item{reference_variable}{name of reference variable}
}
\value{
a vector of the remaining names in the colData
}
\description{
determine what the remaining colData names are when removing the reference
variable.
}
\keyword{internal}
